@@ -6,6 +6,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "start:no-progress": "ng serve --progress=false",
     "build": "ng build",
     "build:ssr": "npm run build -- --app=ssr --output-hashing=media",
     "test": "ng test --sourceMap=true",
@@ -18,24 +19,22 @@
   "private": true,
   "dependencies": {
     "@angular-mdc/web": "^1.0.5",
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/platform-server": "^7.2.15",
-    "@angular/pwa": "^0.13.9",
-    "@angular/router": "^7.2.15",
-    "@angular/service-worker": "^7.2.15",
-    "@nguniversal/module-map-ngfactory-loader": "^7.1.1",
+    "@angular/animations": "^8.2.4",
+    "@angular/cdk": "^8.1.4",
+    "@angular/common": "^8.2.4",
+    "@angular/core": "^8.2.4",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.4",
+    "@angular/material": "^8.1.4",
+    "@angular/platform-browser": "^8.2.4",
+    "@angular/platform-browser-dynamic": "^8.2.4",
+    "@angular/platform-server": "^8.2.4",
+    "@angular/pwa": "^0.803.2",
+    "@angular/router": "^8.2.4",
+    "@angular/service-worker": "^8.2.4",
     "@sillsdev/machine": "^2.1.0-34",
     "angular-file": "^0.5.9",
-    "angular-split": "^2.0.1",
+    "angular-split": "^3.0.2",
     "auth0-js": "^9.10.4",
     "bootstrap": "^4.3.1",
     "bson": "^3.0.2",
@@ -51,7 +50,7 @@
     "material-components-web": "^2.3.0",
     "ng2-charts": "^1.6.0",
     "ngx-avatar": "^3.4.0",
-    "ngx-quill": "^4.8.0",
+    "ngx-quill": "^7.2.0",
     "ot-json0": "^1.1.0",
     "process": "^0.11.10",
     "quill": "^1.3.6",
@@ -59,20 +58,20 @@
     "reconnecting-websocket": "^4.1.10",
     "recordrtc": "^5.5.7",
     "rich-text": "^3.1.0",
-    "rxjs": "6.3.3",
+    "rxjs": "6.5.2",
     "sharedb": "^1.0.0-beta.22",
     "ts-object-path": "^0.1.2",
     "uuid": "^3.3.2",
     "xregexp": "^4.2.4",
-    "zone.js": "^0.8.29",
+    "zone.js": "^0.9.1",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler": "^7.2.15",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "^0.803.2",
+    "@angular/cli": "^8.3.2",
+    "@angular/compiler": "^8.2.4",
+    "@angular/compiler-cli": "^8.2.4",
+    "@angular/language-service": "^8.2.4",
     "@types/auth0-js": "^9.10.2",
     "@types/bson": "^1.0.11",
     "@types/file-saver": "^2.0.1",
@@ -85,7 +84,7 @@
     "@types/uuid": "^3.4.5",
     "@types/xregexp": "^3.0.29",
     "@types/zxcvbn": "^4.4.0",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.1.0",
     "husky": "^2.2.0",
     "jasmine-core": "^3.4.0",
     "jasmine-marbles": "^0.4.1",
@@ -97,13 +96,14 @@
     "karma-jasmine-html-reporter": "^1.4.2",
     "karma-teamcity-reporter": "^1.1.0",
     "lint-staged": "^8.1.6",
+    "node-sass": "^4.12.0",
     "npm": "6.10.1",
     "prettier": "1.15.3",
     "protractor": "^5.4.2",
     "ts-mockito": "^2.3.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.5.3"
   },
   "husky": {
     "hooks": {
