@@ -396,7 +396,7 @@ describe('ScriptureChooserDialog', () => {
     template: '<viewContainerDirective></viewContainerDirective>'
   })
   class ChildViewContainerComponent {
-    @ViewChild(ViewContainerDirective) viewContainer: ViewContainerDirective;
+    @ViewChild(ViewContainerDirective, { static: true }) viewContainer: ViewContainerDirective;
 
     get childViewContainer(): ViewContainerRef {
       return this.viewContainer.viewContainerRef;
