@@ -1,5 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { UserDoc } from 'xforge-common/models/user-doc';
+import { UserProfileDoc } from 'xforge-common/models/user-profile-doc';
 import { ProjectService } from 'xforge-common/project.service';
 import { RealtimeDocTypes } from 'xforge-common/realtime-doc-types';
 import { CommentListDoc } from './models/comment-list-doc';
@@ -9,7 +11,15 @@ import { SFProjectUserConfigDoc } from './models/sf-project-user-config-doc';
 import { TextDoc } from './models/text-doc';
 import { SFProjectService } from './sf-project.service';
 
-const REALTIME_DOC_TYPES = [SFProjectDoc, SFProjectUserConfigDoc, QuestionListDoc, TextDoc, CommentListDoc];
+const REALTIME_DOC_TYPES = [
+  UserDoc,
+  UserProfileDoc,
+  SFProjectDoc,
+  SFProjectUserConfigDoc,
+  QuestionListDoc,
+  TextDoc,
+  CommentListDoc
+];
 
 @NgModule({
   imports: [CommonModule],
