@@ -1,5 +1,6 @@
 using System;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
 using SIL.XForge.Configuration;
 using SIL.XForge.Realtime;
 
@@ -8,12 +9,15 @@ namespace Microsoft.Extensions.DependencyInjection
     public static class RealtimeServiceCollectionExtensions
     {
         public static IServiceCollection AddRealtimeServer(this IServiceCollection services,
-            IConfiguration configuration, Action<RealtimeOptions> configureOptions, bool launchWithDebugging = false)
+            ILoggerFactory loggerFactory, IConfiguration configuration, Action<RealtimeOptions> configureOptions,
+            bool launchWithDebugging = false)
         {
             services.AddNodeServices(options =>
             {
                 options.LaunchWithDebugging = launchWithDebugging;
                 options.WatchFileExtensions = new string[0];
+                options.NodeInstanceOutputLogger = new RealtimeServerLogger(
+                    loggerFactory.CreateLogger("SIL.XForge.Realtime.RealtimeServer"));
             });
 
             services.Configure(configureOptions);
